refactor: extract TranscriptionProvider, TranscribeButton, persist preferences
Architecture cleanup:
- New TranscriptionProvider (React context) manages pipeline lifecycle,
  model loading/unloading, pre-flight checks, metrics collection
- New TranscribeButton component owns all UI: button, GenericMenu dropdown,
  download confirmation modal, transcript display, metrics panel
- useTranscribeAudio hook is now a thin wrapper (~60 lines) around context
- AudioAttachment is back to its original simplicity (~30 lines)
- User preferences (language, model, show metrics) persist via useLocalStorage
  from @rocket.chat/fuselage-hooks (keys: transcribe.language, transcribe.model,
  transcribe.showMetrics)
- TranscriptionProvider mounted in MeteorProvider after DeviceProvider
